CHat tickets

+++ controllers/UsuarioHorController.php
@@ -2,11 +2,16 @@
 
 namespace app\controllers;
 
+use Yii;
+
 use app\models\UsuarioHor;
+use yii\data\ActiveDataProvider;
 use app\models\UsuarioHorSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+
+use app\models\Permiso;
 
 /**
  * UsuarioHorController implements the CRUD actions for UsuarioHor model.
@@ -38,11 +43,34 @@ class UsuarioHorController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new UsuarioHorSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
-
+        if (!Permiso::accion('usuario_hor', 'index')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+    
+        // Obtener el ID del usuario y su rol
+        $userId = Yii::$app->user->identity->id;
+        $userRol = Yii::$app->user->identity->id_rol;
+    
+        // Si el usuario es de rol 4, solo ve sus propios tickets; de lo contrario, ve todos
+        $query = ($userRol == 3) ? UsuarioHor::find()->where(['id_usuario' => $userId]) : UsuarioHor::find();
+    
+        $dataProvider = new ActiveDataProvider([
+            'query' => $query,
+            'pagination' => [
+                'pageSize' => 25
+            ],
+           /* 'sort' => [
+                'defaultOrder' => [
+                    'id' => SORT_DESC,
+                ]
+            ],
+            */
+        ]);
+    
         return $this->render('index', [
-            'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
     }
@@ -55,10 +83,18 @@ class UsuarioHorController extends Controller
      */
     public function actionView($id)
     {
+        if (!Permiso::accion('usuario_hor', 'view')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
     }
+
 
     /**
      * Creates a new UsuarioHor model.
@@ -67,6 +103,13 @@ class UsuarioHorController extends Controller
      */
     public function actionCreate()
     {
+        if (!Permiso::accion('usuario_hor', 'create')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+
         $model = new UsuarioHor();
 
         if ($this->request->isPost) {
@@ -101,6 +144,21 @@ class UsuarioHorController extends Controller
             'model' => $model,
         ]);
     }
+        public function actionUpdateEstatus($id)
+    {
+        if (!Permiso::accion('usuario_hor', 'update-estatus')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+
+        $model = $this->findModel($id);
+        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+        }
+        return $this->redirect(['index']);
+    }
+
 
     /**
      * Deletes an existing UsuarioHor model.
@@ -111,6 +169,13 @@ class UsuarioHorController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Permiso::accion('usuario_hor', 'delete')) {
+            return $this->render('/site/error', [
+                'name' => 'Permiso denegado',
+                'message' => 'No tiene permiso para realizar esta función, verifique con el administrador de sistemas.'
+            ]);
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
